% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{check_data_loaded}
\alias{check_data_loaded}
\title{Check if Data Already Loaded}
\usage{
check_data_loaded(conn, data_source)
}
\arguments{
\item{conn}{Database connection}

\item{data_source}{Path to data file}
}
\value{
Logical
}
\description{
Check if Data Already Loaded
}
