% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_loader.R
\name{expand_env_vars_recursive}
\alias{expand_env_vars_recursive}
\title{Recursively Expand Environment Variables in Config}
\usage{
expand_env_vars_recursive(obj)
}
\arguments{
\item{obj}{Any R object (list, vector, etc.)}
}
\value{
Object with expanded environment variables
}
\description{
Recursively Expand Environment Variables in Config
}
