% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPVdetection-package.R
\docType{package}
\name{IPVdetection-package}
\alias{IPVdetection}
\alias{IPVdetection-package}
\title{IPVdetection: Minimal IPV Detection Using LLM APIs}
\description{
A minimalist approach to detecting intimate partner violence (IPV) indicators in death investigation narratives using Large Language Model APIs. Following Unix philosophy, this package provides simple functions that do one thing well - send text to an LLM and get IPV detection results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xiaosongz/IPV_detection_in_NVDRS}
  \item Report bugs at \url{https://github.com/xiaosongz/IPV_detection_in_NVDRS/issues}
}

}
\author{
\strong{Maintainer}: Xiaosong Zhang \email{xiaosonz@umich.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
