% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_queries.R
\name{read_experiment_log}
\alias{read_experiment_log}
\title{Read Experiment Log}
\usage{
read_experiment_log(experiment_id, log_type = "main")
}
\arguments{
\item{experiment_id}{Experiment ID}

\item{log_type}{Type of log ('main', 'api', 'errors', 'performance')}
}
\value{
Character vector of log lines
}
\description{
Read log file for manual inspection
}
