% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{set_schema_version}
\alias{set_schema_version}
\title{Set schema version}
\usage{
set_schema_version(conn, version)
}
\arguments{
\item{conn}{DBI connection object}

\item{version}{Integer version number}
}
\value{
TRUE if successful
}
\description{
Database-agnostic schema version setting.
Uses PRAGMA for SQLite and metadata table for PostgreSQL.
}
