% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_queries.R
\name{get_experiment_results}
\alias{get_experiment_results}
\title{Get Experiment Results}
\usage{
get_experiment_results(conn, experiment_id)
}
\arguments{
\item{conn}{Database connection}

\item{experiment_id}{Experiment ID}
}
\value{
Tibble with narrative results
}
\description{
Retrieve all narrative-level results for an experiment
}
