% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{load_source_data}
\alias{load_source_data}
\title{Load Source Data into Database}
\usage{
load_source_data(conn, excel_path, force_reload = FALSE)
}
\arguments{
\item{conn}{Database connection}

\item{excel_path}{Path to Excel file}

\item{force_reload}{If TRUE, delete existing data and reload}
}
\value{
Number of narratives loaded
}
\description{
Loads Excel file into source_narratives table for efficient querying
}
